fix: server refactor
