+++ examples/generator-fullstack/server/src/resolvers/resolvers.ts
@@ -1,0 +1,4 @@
+// Placeholder for generated content
+export default {
+
+};
